Predicting Cardiovascular disease risk base on systolic or diastolic blood pressure

Predicting Cardiovascular disease risk base on systolic or diastolic blood pressure

predictcvd(gender, age, Tchol, HDLchol, SBP, DBP, diabetes, smoker, ECG_LVH, t)

Arguments

gender
A binary variable taking 1 if a patient is female and 0 if male

age
A number for age

Tchol
A number for total serum cholesterol in mg/dL measured by the Abell-Kendall method

HDLchol
A number for high-density lipoprotein (HDL) Cholesterol in mg/dL

SBP
A number fo systolic blood pressure in mm Hg

DBP
A number for diastolic blood pressure in mm Hg

diabetes
A binary variable taking 1 if the patient was under treatment with insulin or oral hypoglycemic agents, if casual blood glucose determinations exceeded 140 mg/dL

smoker
A binary variable taking 1 for person who smoked during the past 12 month and 0 therwise

ECG_LVH
A binary variable taking 1 for patients diagnosed with ECG-left ventricular hypertrophy and 0 otherwise (If there is no information, it is 0)

t
A number for time until CVD from the beginning of follow-up

Source

https://www.ahajournals.org/doi/pdf/10.1161/01.CIR.83.1.356 and https://www.sciencedirect.com/science/article/pii/000287039190861B